check file exists before copying
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@6028 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/java/org.wcs.smart.dataentry/src/org/wcs/smart/dataentry/dialog/AssociatedImageInterceptor.java
@@ -217,7 +217,7 @@ public class AssociatedImageInterceptor extends EmptyInterceptor {
 						SmartPlugIn.log("Could not delete cleared file: " + to.toString(), ex); //$NON-NLS-1$
 					}
 				}
-			} else if (!from.equals(to)) {
+			} else if (from.exists() && !from.equals(to)) {
 				try {
 					FileUtils.copyFile(from, to);
 				} catch (IOException e) {
